consolidated `mapLifecycleListener` in `map` subdirectory
